修改删除分类判断
修改删除分类判断

+++ app/Http/Controllers/Admin/FilterWordCategoryController.php
@@ -77,6 +77,10 @@ class FilterWordCategoryController extends Controller
      */
     public function delete(FilterWordCategory $category)
     {
+        if ($category->sensitives()->count()) {
+            return response(['message' => ['该分类下还有敏感词存在，请先删除敏感词再删除分类']], 422);
+        }
+        
         $category->delete();
 
         return response()->json('', 204);
